chatkit_demo/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.2",
    "openai>=2.30.0",
    "openai-chatkit>=1.6.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.38.0",
]
